Modularising the dithered_cube_from_rss code.
I have created a new function, dithered_cubes_from_rss_files which takes the place of the old dithered_cube_from_rss function (this renaming probably not ideal, but the names now better reflect what they actually do.

dithered_cubes_from_rss_file is designed to work on whole FITS files, and takes as input the filename of a list of files and the object names to create cubes for. This function handles opening the relevant files and extracting the IFU object for the corresponding object.

dithered_cube_from_rss takes a list of IFU objects, and does the combining work. All input IFU objects are for the same object.

I have also broken out the guts of the sigma_clipping into a new function. I don't believe the current algorithm is correct.

I have removed much of the plotting code, since modularising was going to break some of it, and I thought it best to just take it out rather than keep rewriting it while figuring this all out.