Modified utils.mad to apply along an axis as well as on the whole array.
Since mad includes an "axis=0" argument, I naively assumed it could be told to apply the mad only along an axis. But I was wrong: the keyword is ignored.

I have fixed this up to where axis=None (the default) still behaves as before, computing the mad of the whole array. However setting axis=n now computes the mad along the nth axis, returning an array of one fewer dimensions.