sigma clipping now includes weights.
I have modified the cubing code to now account for the weights of the fibre-to-grid mapping before trying to do the sigma clipping. Additionally, I changed the sigma clipping to be done on the real data, not normalised data. These two changes now allow the cubing code to generally pass the "hot_pixel_clipping" test.

With the normalisation as advised by Rob, the null test clips considerably more pixels (with out it, there are only a few pixels clipped). Unfortunately, I can't follow Rob's reasoning in the paper, although it is probably correct and I just can't understand why. Anyway, I have turned it off for now.

One final note: currently the null test just demands that the number of clipped pixels be less than 100, but this is in fact not correct. The actual number of pixels to be clipped even in the null case could be predicted by statistics, but that has not been done here yet.