Converted nanmedian and nansum to bottleneck. Added file with cubing_tests.
Bottleneck provides factor of 2x speed up in numpy operations involving nans. I have made the change in such a way that the software reverts to the standard numpy.nansum and scipy.stats.nanmedian gracefully if bottleneck is not installed.

This makes a difference in the speed of the cubing code, which spends almost all of it's time in these two key functions. For me, this improved the speed of sami.general.cubing_tests.speed_test() by a factor of 5.5x!!!

Also, I stupidly had not added cubing_tests.py to mercurial, and finally managed to delete it. I have recovered/recreated it, and now checked it in. Sorry!