Reviewed sigma clipping after long discussion with Rob Sharp. There are still problems.
I had a long phone call with Rob, and I understand now the need to normalise the spectra before trying to do the sigma clipping. I changed the code back to using this normalise data to compute the sigma clipping mask (once outliers are clipped, the original, unnormalised data is used to compute the sum).

While I was at it, I realised another error. The original code incorrectly used the standard deviation as the variance when doing the sigma clipping. Changing this to be a variance (by squaring the result of the MAD estimate of sigma), makes the sigma clipping catch a stupid number of pixels in the tests, and fails them both.

This may seem counter-intuitive, but sigma is being measured as less than 1, so squaring it makes it even smaller, making the constraint tighter. The input test data has sigma=1, so the fact that the sigma used for clipping is typically around 0.001 or smaller is a real problem. I'm not sure what's wrong, and I don't want to try to figure it out today, so I've left it as it was, @TODO'd it and moved on.