Moved the drizzling into the main slice loop of cubing, and modularised.
I renamed the fibre_overlap_map class to SAMIDrizzler, since we seem to have a lot of overlap named things floating around.

I created a new member function of the SAMIDrizzler class which includes what was previously a loop within dithered_cube_from_rss, namely the loop which creates the "fibre overlap maps" for each fibre across all the observations to be combined. This function is now called within the main loop over slices (in anticipation of it becoming DAR aware.)

Finally, the new function caches it's last arguments, so if the request is the same, the drizzle is not recomputed, meaning that there is no performance hit if there is no DAR.