utils has been broken into several modules.
ifu.py - which contains the IFU class
circ.py - Jon Neilson's circle/drop code
other.py - the other small functions we're using

The reasoning is two fold: there are a few things which need to be added to ifu.py to handle RA and DECs for the DAR code (I could put them in the DAR code, but that seems silly given they will be fairly general things), and secondly, I'm going to need to do some optimisation to the Jon's code in order to make the the DAR run in a reasonable amount of time. Optimisation will be nearly impossible if these functions aren't in their own module.

It would also make sense to put the fitting stuff inside here as well, but I'll leave that to another day.

I have written a __init__.py file which maps this out without any change to the user interface, and made sure it doesn't break anything else.