"Corrected" sigma clipping.
The sigma clipping now correctly clips on 5-sigma, rather than 5-variance, as it had done previously. In the original implementation, the standard deviation had been mistakenly passed as the variance.

However, this now does not pass the sigma clipping tests. This seems to be because the MAD is sometimes far to small for a set of pixels at a point in the output grid. Keep in mind that there are typically only a few ( <= n_obs) pixels to compute the MAD of, so sometimes the value is artificially low. The result seems to be that this clips far more pixels than it should. Fixing this will require determining the variance/stddev via alternative methods. The most likely would be to use the 2dfdr variance corresponding to the median data value of the pixels to be combined (Rob's suggestion some time ago). This will take some time.

Unfortunately, I'm not convinced that the sigma clipping approach we have implemented here is physically valid, and now suspect that it will be just as likely to clip real data as bad values, but I haven't finished working through the numbers. So that will have to be done first, before the sigma clipping can be finished.