DAR functions and structure in place, but no correction.
All of the code to compute the DAR correction and apply it is in place. I have largely taken the work Mike Childress did and confirmed that it gives the correct answer, and then added it as a new module under the utilities package.

Inside the cubing code, a few things have been rearranged so the fibre position information can be updated by the DAR correction at each wavelength step (this was in an earlier commit). There is a new class, DARCorrector, which is the class which will contain the functions to do the DAR correction via various methods, including turning it off altogether.

Unfortunately, this is now as much as can be done without propagating the header information, so that is now my highest priority. Currently the DAR correction method is set to 'none' and the cubing code should generate fully valid cubes.

Once header propagation is in place, the next step will be to optimise the speed: the current implementation will take hours, possibly days, to produce each cube.