First go at header keyword propogation.
The sami.utils.ifu.IFU has been modified to carry around most (all?) the relevant meta-data, which it previously just discarded. A few other minor cleanups there.

The cubing now takes the information from the input Primary Headers and puts it in a meta-data table extension of the output file. The extension is a binary FITS table with n_rows = number of input files, and columns for each header keyword in those input files. The scope is there to add most other meta data from the input files to that table.

This version works (produces a valid FITS output with the metadata table constructed), but I have commented out the actual work of producing a data cube for testing (output data is just zeros.)