Simple DAR Correction in place and tested, not yet correct.
I have written a test for the DAR, which basically takes a few slices from real data, runs it through the cubing code, and then fits the position of the peak in the output data. This is cubing_tests.dar_correction_test().

There was also a problem with units (I mistakenly thought Mikes ADR functions used microns, but they use angstroms.) Other fix is to convert the correction to microns, since the fibre position information is in microns (this should probably be cleaned up).

Currently, the correction is not big enough, and not in the right direction. However, there is some 90-degree rotation in the cube which is not understood (see comment in the code from LF just before cubes are written out). This will need to be tracked down. I'm less certain why the magnitude still seems to small (by factor of 2-3) since the original tests showed it gave the right answer.