Implemented Sam's WCS calculation into cube header creation
Implemented Sam's WCS calculation, based on cross-correlation of SDSS g-band images with a flattened blue cube. This is mainly contained in the two functions, getSDSSimage and WCS_position. Also defined a couple of global variables describing the grid parameters for cube creation (though passing these to each function as local variables would be better), as well as some minor changes to get the create_primary_header function to play nicely with the WCS_position output.

Also slightly modified utils/ifu.py to include the group RA and DEC as properties of the ifu class.