Changed the telluric.py file to solve for the telluric correction factor and saves it to the red frame in the extension "FLUX_CALIBRATION".
The telluric correction factor (as a function of wavelength) is solved by fitting a linear slope to the continuum either side of the telluric features, forcing all values outside of the telluric regions to be the value of the slope at the respective wavelength, then diving the entire spectrum by the linear fit. This produces the correction factor that is mostly unity, but non-unity in the telluric regions. This is saved to the red frame as the 3rd row in the extension "FLUX_CALIBRATION". This correction factor is then to be multiplied with the object data to achieve telluric corrected data.