WCS handles missing imaging data
Modified the WCS routine so that if the SDSS image server returns a blank fits image (i.e. the requested image coordiantes are outside the survey area) it will use the nominal catalogue position of the galaxy in place of an accurate cross-correlation determination.

Also added an extra header keyword, WCS_SRC, to record which method was used to determine the WCS header parameters.