Standardised test data location.
The test data locations were specific to my computer. Now I have set the code to look in a "test_data" directory which is expected to be in the same sub-directory as the sami package itself. We will have to decide how to maintain this data in the long run (i.e., check it into mercurial or distribute it separately).