Found and fixed the 90-degree rotation problem in the cubing code.
There was a crypic comment in the cubing code: "90-degree rotation that I can't track down." This made troubleshooting the directions for the DAR correction troublesome, so I have found and fixed the problem. It turns out the circ,py routines assume y is in the first index and x is in the second index. This means that when one bindly does imshow(output_frac_map), the image appears to be the correct orientation, but in reality, imshow must have it's input array transposed so the coordinates match the actual indexing. I have carefully documented the resample_circle behaviour in it's doc-string.

Then, the cubing code immediately transposes the result when it calls circ.resample_circle, and therefore the strange transposition of axes when creating the FITS file in dithered_cubes_from_rss_files is no longer needed and has been removed.

Also cleaned up some of the comments and formatting in cubing.py.