Simple DAR correction in place and working (with caveats).
DAR correction is now applied using a theoretical correction derived from atmospheric values stored in the input FITS header. The correction is applied slice-by-slice as the data is mapped from input RSS files to output grid. this imposes a factor of ~11 speed penalty in the cubing (cubing with DAR: whithout sigma clipping, about 4 minutes; with sigma clipping, approaching 30 minutes).

As I write this, I realise that currently the DAR correction appropriate for the first RSS file in the input list is blindly applied to all the RSS files contributing to the cube, which is incorrect and will need to be fixed.

There is also a remaining "flip" along the x-axis of the data which is not explainable (currently, it's flipped in such a way as to make the data look right, but also made determining the direction to apply the DAR correction hard.). compute_parallactic_angle returns the correct angle to the zenith, and the flip is accounted for by changing the sign of the x-component of the correction before applying to the fibre positions.