Continuing changes to data-import routines.
Importing of cubes and RSS strips has been implemented. Fibre table not yet taken care of, its mix of datatypes require insertion as a compound dataset. Header information is being recorded, but comments are not supported by HDF5. This still needs to be sorted. And there are currently no variance headers accessible through the IFU object. I will return to this on Monday.