Introduced eat_cube() function for standardising cube import.
Wrote a function to import all extensions of a cube in a homogeneous fashion. Minimises error margin and makes maintenance clearer. Could not extend use of this function to input of RSS strips, as they are accessed through the IFU object, not as HDUs. Verbose input still in place for RSS. Will look at ameliorating this situation.