Automatic identification of observation type. Plus updates.
Added a couple of lines in import_cube() to diagnose the observation type through the ID number. GAMA IDs are up to 7 digits; F-stars are always 8 digits and start with 1000; cluster IDs are 10 digits long and start with 9. Tested, it works.

Also updated import_many to conform with new import_cube input arguments.