Modularised fetch_cube().
Wrote a series of functions that the export() code can envoke to perform small tasks:
 + checkSAMIformat() checks that the input file is SAMI-formatted.
 + getObstype determines observation type.
 + getVersion determines the latest version of a target available, if the provided 'version' is a blank string.
 + getTargetGroup() finds the h5 group where data for the requested target live.
 + completeCube() checks if the archive holds all [data, var, wht] extensions fot the requested target
 + defOutfile() defines the output (FITS) filename, if one is not provided.

WARNING: This version of the code contains some collateral in export() from its upcoming adaptation to the new filesystem. This will be the only export code and the new fetchCube will just eenvoke export() with pre-set arguments.