Remove dependence on IRAF from the offset computation code, align_micron.py.
Wrote two new functions which are the model that is used to fit to the plate to determine the offsets, and a related function which just returns the normalised residuals for passing to the leastsq optimisation function.

Have commented out the IRAF routines, and replace with a leastsq fit, and then a recomputation of the model points using the results of the leastsq fit. This is all put back into the appropriate data structures of the original code.

In theory this should return the same result as the previous changeset, but I have no way to test since I cannot get pyraf running on my machine. A visual check of the results appears to give sensible values.

Final change is that this version of find_offsets returns the results of each leastsq as a python list of tuples.