Updating import_table() to add tables to 'SAMI/<version>/Tables'.
Code renamed to importMaster(), as it is specific to a GAMA-based SAMI target table. Tables will now live in a 'SAMI/<version>/Tables' group, so they can be version controlled. the 'h5dir' argument has been removed from the calling sequence, and the most recent version will be chosen by default (a manually inpu argument for now). Tables will have to be included manually, that is, with every data release there will be the manual addition of a MASTER table and any other updated HLSP/DMU tables. If the table has not changed, then it should be linked to the previous version, not duplicated. The name of any compound dataset holding a DMU should include the DMU version. That way we can support multiple versions of a single DMU in the same Data Release.

Additionally, due to a limitation of PyTables (inability to interpret attributes with empty value fields), I have restricted the inclusion of the temporary [comm] attributes to only those header items with comment fields in import_cube(). Blank header tickets themselves are replaced with '[blank]'.

Finally, I have introduced a little version-checker loop in importMaster(), as I was not able to easily adapt sami.db.export.getVersion() to interpret the contents of an h5 file with either h5py or PyTables. Incidentally, getVersion() should probably be moved to utils.py.

TODO: Need to introduce a version branding system for these tables, including in the name. That will be in the next commit though, enough updates here!