Adapting query.py to new filesystem.
Started with query_master(), which has ben renamed to queryMaster():
1) Adapted to change of location for DMUs and SAMI target tables.
2) Changed h5file to hdf and h5in to h5file for consistency with ingest.py.
3) Now calling getVersion() within queryMAster(). Added version arg.

Point (2) above also applied to query_multiple, which is now queryMultiple().

NOTE: Once versioning is applied to tables, the query codes will break. We will need to soft-link the latest version of each table to its name without the version suffix, to make this simpler. Changing the query codes to add the table version would be bad book-keeeping.

NOTE: There were no changes to 'ingest.py', not sure why it was picked up. Will do a quick diff and post a message if necessary.