DAR (spectroscopic) Correction in place and working.
The way the DAR correction is implemented within cubing.py has been changed. Now a single function is called which updates the fibre position information (which is now an array over wavelength) before the drizzling per wavelength slice is started. The drizzling then updates at every wavelength slice. Alternately, if the drizzle is not to be updated at every wavelength slice, then this would be incorporated into the drizzling, not the DAR.

There is a corresponding test for the DAR correction in cubing_tests.py, and the code passes it. In order to make this work, I've created a diagnostics module which is currently just a dummy module to store diagnostic data (which is much cleaner than the approach I took for the sigma clipping tests.) This should be expanded further in a separate branch.

mc_adr.py has some changes to account for moving the parallactic angle computation into the DAR Corrector class (which is much cleaner, and allows it to return dar correction components in the explicit east-west and north-south directions.