Cubing code now respects drop size.
It seems that the drop size reduction factor was never correctly implemented in the code. A flag was added to the various functions in a cryptically described commit from a long time ago, but was never fully implemented. (So the drop reduction has never been in the repository).

I have created/modified a set of module level variables which control the drop size and output grid dimensions and pixel scale. For someone wishing to change these, the best approach would be to change the values after the code has been imported, but before calling any of the cubing commands or creating a SAMIDrizzler class.