DAR correction now uses correct declination and zenith distance for bundle.
There is a new function, zenith_distance, in mc_adr.py, which gives the zenith distance to an object.

The IFU class now has two extra data members, "ra" and "dec", which contain the actual RA and DEC of the bundle. Note these are not exact (in the same way that "meanra" and "meandec" are not exact for the centre of the field plate.)

In cubing.py, the dar_correct function now computes the zenith_distance as the time averaged mean of the zenith distance over the observation by integrating the zenith_distance function. Also, the declination is now set to the actual declination of the bundle using the new data members of IFU described above.