Added various logging messages and made some routines more robust.
I still think align_micron will fail if the fitting for a particular IFU fails for any reason (where by fail, I mean raises an exception). I've written a try/except clause for this, but decided to just pass the exception through as the old code does (a debug message is now printed).