Introduced functions that copy h5 groups/objects through Python.
Working toward an export() code that will quickly create a 'baby' HDF5 block wither to export, or as an intermediate file for FITS packaging. The code is copyH5() and it is output format specific. In FITS mode it creates a temporary file (named after a random string, which contains not hard, but linked data. This makes it very fast and very small, so gentle on the staging drive. In HDF5 mode it creates a named file that is to be staged.

Other changes:
-- Edited getTargetGroup() to return both g_target and obstype. There is no need for the duplication there.
-- Now need to phase out getObstype() in fetch_cube() and export().
-- export() can now be modified to evoke copyH5() and then unpack fits files.
-- Obviously the fits unpacker script needs to be written.