Changed flux propagation through cubing code.
This now (I believe) reflects the description in drop_size.pdf. (The regular description, not the Alternate Convention.) There are a few variable names which should probably be reconsidered now.

I don't know if this is the final answer, as the values in the weight map with this convention are somewhat arbitrary. This was the cleanest and fastest for me to implement, but an approach where the weight map has some more intuitive physical meaning might be better (probably something between the two versions proposed in drop_size.pdf, unfortunately).