Added a function to DARCorrector to set itself up for a given IFU object.
There are really two additions here. The first is a new function in the DARCorrector class which takes an IFU object as an argument, and then sets up the current DARCorrector instance for the parameters of that IFU object. Old method still works fine.

Second is the addition of the "wavelength" property. When set, this property also updates the dar_r, dar_east, and dar_north values.

This is all new, and does not in any way affect the old interface. However, it would be a good idea to simplify the usage in cubing.py.