Fixed DAR correction + alignment problem.
The DAR and alignment sections of the code now appear to be working together properly. The alignment is now done on the red arm of the data only, and the results are copied across to the blue arm. The alignment applies a DAR offset to each galaxy before doing the fitting. All DAR caculations are now based on the fit to the star in the telluric correction stage - this ignores the change in DAR magnitude across the plate, but should be good enough.

Other changes:

A bunch of IO functions were moved from dr/fluxcal2.py into utils/fluxcal2_io.py to avoid a circular import.

A new do_dar_correct keyword added to get_centroid, so DAR correction can be swtiched off.

In the manager, cubing has been split into two stages: measure_offsets and cube.

The IFU class reads in atmospheric parameters saved by fluxcal2, if available.