Changes to the logging system to make it more useable.
I've set up the logger to where it now will log all messages of DEBUG or lower to the file given by the environment variable SAMI_LOG_FILE. If this is not set, then it only logs to the screen.

This is set up in the 'sami' logger. All other loggers should be children of this (assuming they are initialised with "logging.getLogger(__name__)", and therefore should inherit these settings. Log levels for individual modules can then be reset.