Added a bunch of functions to the sdss module.
These were developed a long time ago, but I just never got round to adding them to the SDSS module. There is a plotting tool that overlays a SAMI bundle in real sky coordinates to any fits file or a metadata-enabled jpeg. There is a g-band mock image production tool (RSS and cube flavours). And there are some utils to make all this work.