Initial version of gpuCirc.py.
The GPU accelerated version here differs from circ.py in two key ways. First, the algorithm for computing the drizzling is different: it uses many rays to determine the approximate drop coverage of each pixel (accurate to about 0.2% per pixel). Second, the core algorithm for each pixel is executed using OpenCL, which will use the GPU if available, or all of the cores/threads of the CPU otherwise.

This version requires PyOpenCL, and a few tests suggest it is a drop in replacement for circ.py (with the caveat about some small change in accuracy as described above).