Fixed a few more issues with KOALA Cubing, including changing field size.
For the change in field size, I've decided to have the plate scale and fibre diameter be a property of the IFU object. I have not updated the SAMI IFU class to reflect this. The numbers in config.py are now ignored for cubing.

Also added the missing slogging.py module.