Seems that even after Tony's changes, the KOALA fibre table is not correct. This update should correct that.
The issue seems to be that Tony's code computes the PA as the sum of TEL_PA, and INST_ROT. For reasons that are unclear, INST_ROT was 90? for data from 2014-Oct, variable in his test data (as he said this was easier to change than the TEL_PA), and -90? for data from 2015-July.

Really not sure why the INST_ROT is changing, but its making a mess. This code will now issue an error if the INST_ROT is not as expected.