Added code to the flux-standard identification to correct for cases where an alternate aperture has been used.
If the telescope aperture is set to A or B (instead of REF), presumably because APOFF values have been entered so the telescope will aquire an object to a guide bundle rather than the centre of the plate, then the RA and DEC information in the file will be offset by this amount. This offset can (does) cause the standard-star identification to fail. This modification resets the RA and DEC by reversing the effect of the APOFF on those values.

Added to correct/handle data taken 3 March, 2016.