Adjusted the signature of `reduce_all` to handle 3am syndrom more easily.
`start` and `finish` are now the first two keyword arguments, so that if one forgets to put the keywords in the call, it still behaves as one might expect. Also now validates the values of start and finish and displays a useful message to the user if they are not valid.