Fixed a problem in the manager where combine_lflat/dark/bias is called, but the corresponding files don't exist.
Previously, if `combine_bias/dark/lflat` was called when there were no corresponding files (regardless of whether `reduce_bias/dark/lflat` had been called), the package would make a link to a non-existant file. Subsequent invocations would find this file, but os.path.exists returns False (broken link), but os.symlink errors because there is already a symlink in place.

This corrects the logic such that (a) the broken symlink will not be created, and (b) if the symlink is there but broken and `overwrite=True`, then the existing link will be removed.