Added masks, scaling of fluxes for aperture sizes, and fudge factor for area mis-match.
Previously, the aperture flux was given as a flux density per spatial pixel (I think: based on the binning code and what I assume the cube units are...) Now the aperture flux is scaled up by the number of spaxels included (so larger apertures will have larger fluxes).

Also, I've added a fudge factor to the header which gives the ratio of the actual area (in whole spaxels) vs the expected area of the elliptical aperture.

Finally, the mask of pixels included in the aperture is also written as another extension in the output FITS file.