Enable reduction of twilight sky frames (MFSKY) to make tramline maps.
Add a keyword use_twilight to make_tlm() that enables twilight frames to be used to generate tramline maps (and also sigma maps).  If use_twilight is True the code will make copies of any MFSKY files and change the class of the copies to MFFFF.  The list of new files is then reduced to make a TLM, just as other MFFFF files are.  The new copies are reduced in a separate call to reduce_file_iterable() as they are not in the full file list in the same way that original files are.  A related change needed was to modify copy_as() so that it adds the correct attributes to the FITSfile object made in the copy.  In particular, tlm_path is needed and a call to set_check_data() was also added.