Bug fix: Make sure we always use the TLM from the same file if reducing an MFFFF
When 2dfdr is processing an MFFFF it will always write a new TLM file, even if we pass a different TLM file name to the code.  As a result, the new TLM made will overwrite the file that has been passed to 2dfdr.  We don?t want that to happen as, for example, a twilight TLM file could be overwritten by a MFFFF TLM.