Add extra function to FITSFile class to get the detector name, e.g. E2V2A, and change reduction options so that new CCDs do not use darks or biases.
The new CCDs (that were installed in SAMI a couple of years ago) have better characteristics and so we do not need to use darks or biases with them.  So we add a function to the FITSFile class to get the FITS keyword DETECTOR.  This is then tested in the tdfdr_options function and bias/dark subtraction is switched off for E2V2A and E2V3A (the names with A at the end are the new CCDs).