Extensive cleanup and refactor of tdfdr.py
tdfdr.py was looking a bit sad, so I gave it some TLC. In particular, GUI execution is now handled by a different function than pipeline (aaorun) execution.

Also, I?ve greatly simplified the handling of the IMP_SCRATCH variable and temporary directories. Instead of changing the current working directory of Python all the time, this instead uses the ?cwd? keyword of the call to subprocess to start the sub process in the correct current working directory. This avoids any risk that the cwd could be changed inadvertently by conflicting actions running in co-routines (and is just cleaner anyway!). In the same vein, I eliminated the constant changing of the environment variables to handle the IMP_SCRATCH, again, instead choosing to pass the exact environment to subprocess directly.

These changes allowed the context managers that had handled this to be eliminated. Now the only context manager handles creating the lock files and then removing them later (I note that I?m unsure that this approach is really thread-safe. I should grab the version used in FIDIA, which I spent a lot of time ensuring was thread safe.)

This refactor also lead to some minor changes to the manager to update it?s calls to tdfdr functions. Finally, the manager still used ?raw_input? in many places, which is not compatible with Python 3, so I?ve changed this to ?input? and added a six.moves import to make this backwards compatible.