Fixed another issue with parallized manager loading that caused duplicated files in the manager.
It seems to me (without thorough testing, mind you) that the manager import was importing the correct raw files only by happenstance, not by design. It os.walkdir?s over the manager directory, searching for files that look like 15feb10001.fits. This search will find both the original files (in the raw sub-directory) and any links to those files elsewhere in the tree. The duplicates were removed by checking if they already existed in the manager. Only the files in raw were kept because os.walkdir by default does a breadth-first traversal, so it would find ?raw? before it would find ?reduced/180215/Ysadf?? and therefore the raw would be added, and subsequent ones rejected. This seems pretty dangerous reasoning, and it would be more sensible for the manager to not consider files other than those in the raw sub-directory, so I?ve changed it to reflect that.

Nic and I can?t think of any reason that the manager needs to know of files outside of the raw directory (all other files are basically found by mapping from the raw files to their various reduced counterparts).