Cleaned up type problems in offsethexa that were leading to incorrect behaviour and warnings.
Ugg. CSV files without everything plugged will have columns with empty strings in them for the probe number. CSV tries to coerce these to ints, and succeeds if all bundles are listed. If not, then it will remain a string. This means that offsethexa doesn?t know if it is going to get a string or an int. It?s been coded assuming it gets a string, and this then causes errors and strange behviour.