Not all checks were being added when the manager was run with `n_cpu > 1`, now fixed (hopefully).
I had already realised that the 2dfdr checks were not being run earlier, hence the attempt to set up the code to run using co-routines instead of multiprocessing. However, I hadn't appreciated that some of the other checks were not being updated either.

To overcome this, I've updated the code such that the checks are never updated in a parallel processing step, instead, the checks are updated in the single thread that is calling the parallel processing either before (or more likely after) the parallel processing has completed and returned.

This fix eliminates the original need for co-routines, so I think I can safely move this back to the non-co-routine based version, and then merge it back into the regular py3 branch, which is intended to be backwards compatible.