Closing cross-correlation alignment branch per Francesco's arguments (below).
TL;DR In my view, the cross-correlation method does not work and will not work.

As I mentioned earlier to Nic, the reason is that the cross-correlation relies on the presence of multiple features. As such, it has (much) better accuracy than the default method whenever there are multiple sources or resolved galaxy structures in the IFU. Unfortunately, for most of the SAMI galaxies, the net result is a loss of precision by a factor of three (RMS in Delta(R.A.) and Delta(Dec) rises from 0.08" with the default method to 0.24" with the cross correlation). Naturally with enough effort the method could be made to work, but I do not have time now or in the near future. Maybe for Hector, when we will have too many galaxies to inspect visually like Sree did?).