Update to calculate the effective airmass of an observation
As part of the IFU class calculate the effective airmass (this is in
utils/ifu.py).  This replaces the previous version that is calculated in
fluxcal2.py, but that only used the ZDSTART keyword.  Now the code takes
the coordinates of the field, as well as the location of the observatory
and the time/date of observation and uses astropy routines to calculate
the ZD.  Changes made mostly to utils/ifu.py and dr/fluxcal2.py