Update to effecrive airmass calc, to move the calulation to a different location in the code.
Fix to new effecive airmass calc to put this in the fluxcal2.py routine only
and not have it in the IFU class.  Instead the function in fluxcal2 which is
called calc_eff_airmass() does this, and just takes the a header as input and is
called by the two rotuines that do extinction correction, remove_atmosphere() and
remove_atmosphere_rss().