Update transfer fucntion combining to combine all standards from a run and other minor changes.
The flux calibration files arfe not only grouped by ccd, not name, date or field.  They are
averaged together, but we also add a S/N weightng to the combination (weight by (S/N)**2) so
that frames that are too low in S/N are not given too much weight in the fit.  Some tests
show that they tend to diverge from other solutions when the flux counts are below ~500-600
in the blue arm (median).  Finally also allow spline fit to do some extrapolation, so that
we can calibrate the full range of all the bundles.  The extrapolation is low order and
seems robust (does not diverge crazyily at the very end of the spectrum).