Addition of routines to allow flux calibration based on secondary standard stars.
A new top level function in manager.py is added fluxcal_secondary().  This does
a number of steps.  First it fits the extracted secondary spectrum to a set of
theoretical templates.  Then the best fit template is compared between different
frames of the same field and the weights and templates are averaged across all
frames.  The optimal template is made from the 4 closest in teff and [Fe/H] to
the best fit.  Secondary TFs are derived based on the best fit model.  Allowance
is made for a galactic dust model as the stars should be halo stars and above
the dust.  Finally the TF is applied, but this could be invidually or based on
the mean over the field (adjusted for overall normalization).

Requirements: ppxf, dust maps in dust/, kurucz model files in standards/

There are also a small number of other minor changes.  For example, allowing
generate_subgrid() for integrating over fibres to weight by a fibre profile.
This is switched off as a default in the code currently.